Factored out Applications specific handlers to a separate submodule
